Refactor: Manager: Move the ansible-related code to a separate manager class
